pyproject.toml
[tool.poetry]
name = "codementorapi"
version = "0.2.0"
description = "Python client for the Codementor API"
authors = ["Zack Plauché <zackknowspython@gmail.com>"]
readme = "README.md"
repository = "https://github.com/yourusername/codementor-api"
documentation = "https://github.com/yourusername/codementor-api#readme"
keywords = ["codementor", "api", "client"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
playwright = "^1.49.0"
pydantic = "^2.10.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
